+++ paddle/fluid/operators/math/beam_search.cc
@@ -65,7 +65,7 @@ class BeamSearchFunctor<platform::CPUDeviceContext, T> {
     auto *selected_scores_data =
         selected_scores->mutable_data<float>(platform::CPUPlace());
     auto *parent_idx_data =
-        parent_idx ? parent_idx->mutable_data<int>(platform::CPUPlace())
+        parent_idx ? parent_idx->mutable_data<int64_t>(platform::CPUPlace())
                    : nullptr;
 
     // fill in data
@@ -75,7 +75,8 @@ class BeamSearchFunctor<platform::CPUDeviceContext, T> {
       low_level.push_back(low_offset);
       for (auto &item : items) {
         if (parent_idx) {
-          parent_idx_data[low_offset] = static_cast<int>(low_level.size() - 1);
+          parent_idx_data[low_offset] =
+              static_cast<int64_t>(low_level.size() - 1);
         }
         selected_ids_data[low_offset] = item.id;
         selected_scores_data[low_offset] = item.score;
